+++ src/main/java/org/homio/app/model/entity/SettingEntity.java
@@ -5,6 +5,8 @@ import jakarta.persistence.Column;
 import jakarta.persistence.Convert;
 import jakarta.persistence.Entity;
 import jakarta.persistence.Transient;
+import java.util.Collection;
+import java.util.Set;
 import lombok.Getter;
 import lombok.Setter;
 import lombok.experimental.Accessors;
@@ -21,9 +23,6 @@ import org.homio.app.manager.common.impl.EntityContextSettingImpl;
 import org.homio.app.repository.SettingRepository;
 import org.jetbrains.annotations.NotNull;
 import org.json.JSONObject;
-
-import java.util.Collection;
-import java.util.Set;
 
 @Getter
 @Setter
@@ -144,7 +143,7 @@ public class SettingEntity extends BaseEntity {
     }
 
     @Override
-    protected int getChildEntityHashCode() {
+    protected long getChildEntityHashCode() {
         return 0;
     }
