+++ src/main/java/org/homio/app/model/entity/widget/WidgetTabEntity.java
@@ -1,7 +1,12 @@
 package org.homio.app.model.entity.widget;
 
 import com.fasterxml.jackson.annotation.JsonIgnore;
-import jakarta.persistence.*;
+import jakarta.persistence.Column;
+import jakarta.persistence.Convert;
+import jakarta.persistence.Entity;
+import jakarta.persistence.FetchType;
+import jakarta.persistence.OneToMany;
+import java.util.Set;
 import lombok.Getter;
 import lombok.Setter;
 import org.homio.api.EntityContext;
@@ -14,8 +19,6 @@ import org.homio.api.model.JSON;
 import org.homio.api.util.Lang;
 import org.homio.app.manager.common.EntityContextImpl;
 import org.jetbrains.annotations.NotNull;
-
-import java.util.Set;
 
 @Getter
 @Setter
@@ -67,7 +70,7 @@ public final class WidgetTabEntity extends BaseEntity implements HasOrder {
     }
 
     @Override
-    protected int getChildEntityHashCode() {
+    protected long getChildEntityHashCode() {
         return 0;
     }
